Hello,
Debian appears to have removed the version of discover-data that
systemimager uses from their FTP site. The timestamp on the FTP folder
indicates that it was probably done on the 19th. Here is a patch that
updates systemimager to use the next available version.

--
Josh Aas
Linux System Software
Silicon Graphics, Inc. (SGI)