- fix: resolved a bad issue with bittorrent transport.
If clients are not able to reach internet the installation can hang
with BitTorrent transport. By default the --upnp option is used to
enable automatic port mapping (for peers behind NAT routers or
firewalls).

With this option the following check is done by bittorrent-console (see
BitTorrent-4.4.0/BitTorrent/NatTraversal.py near line #72):

    try:
        s = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
        s.connect(("bittorrent.com", 80))

In some configuration (in particular if clients cannot reach internet)
the connection hangs, blocking the whole installation process, since
there's not a timeout in the connect.

The fix consists of using --no_upnp option with bittorrent-console in
the clients.