- allow to put local.cfg and ssh keys directly into the initrd.
The trick is to add the needed files into the initrd_template and use
si_mkbootpackage to re-generate the initrd (also when using BOEL).

In this way it's possible to use ssh keys and local.cfg when
autoinstalling via a CD, or an USB pen and also via network (in the
cases that the boot is done in a secure network and the image is
downloaded over an insecure network - i.e. Internet).