- fix: SSH transport didn't work in server-driven way.
Running the autoinstall script with si_pushupdate is not correct! The
hardware detection and pre-install scripts steps were skipped.

Now si_pushupdate opens a SSH tunnel from the image server to the
clients, in the opposite way of the client-driven approach (where
clients open the SSH tunnel to the image server).