-fix: sort -u doesn't work to get a unique list for pre and post install scripts.
The command uniq (as sort -u) works only if the list is sorted, but we
want to preserve the order of the different classes (IMAGENAME,
BASE_HOSTNAME, HOSTNAME, all, OVERRIDES), so we cannot sort the whole
list.

The solution is to use a custom unique() function to remove duplicate
entries preseriving the original order.