- added si_pushoverrides command
This command can be used to push files from the image server to the
clients using the overrides. It uses si_pcp (rsync over SSH) to
distribute files to the clients. It is also possible to define node
groups and cluster topology using the configuration file
/etc/systemimager/cluster.xml.

The approach is totally server-driven. Typically this command can be
used to distribute configuration files to the clients in a secure way,
because the clients don't need to access directly to the image server.