- replace discover with pci-automod in the initrd
pci-automod is a bit slower than discover, but it has the big advantage
that it doesn't need a dedicated DB to be maintained (like
discover-data). It is fully based on the informations found in sysfs and
the modules.pcimap file (shipped with kernel modules).

Unfortunately we can't still remove discover, becuase some old
kernels don't support sysfs, but when it's possible is always better to
use it in order to have a more reliable hardware support.