- fix: create special files into the initrd_template in a way compliant with the   Debian packaging policies.
It's forbidden to ship special files in a deb package. The correct procedure is
to create them by a post-installation script and using the MAKEDEV command. 

The post-installation script is created at compile time in debian/rules, using
the input file:

debian/systemimager-boot-arch-standard.postinst.in

This is necessary to properly rename it as the systemimager-boot-<arch>-standard
package, since the build architecture can be only evaulated at compile time.