- introduce the concept of host group
Allow SystemImager commands to be aware of host groups defined in
/etc/systemimager/cluster.xml. Host groups allow to consistently speed-up the
typical management operations on your clients, specifying a single name (the
group name), instead of remembering the complex host ranges of similar clients
that work in the same logical group. This also reduce the risk of errors due to
accidental oversights/omissions of hosts.