- implemented host group concept for autoinstall script and pre/post install   scripts
This is a complete restyling of how autoinstall and pre/post install scripts are
chosen by the imaging clients.

The new command si_clusterconfig has been introduced to edit the configuration
of the groups of your clients (/etc/systemimager/cluster.xml). The command also
provides to create a cache of the XML configuration in a plain text file
(/var/lib/systemimager/scripts/cluster.txt), easily parsable from the minimal
busybox environment of the imaging clients.

Now the autoinstall script is chosen using the following algorithm:
- if SCRIPTNAME is explicitly specified by installation parameter, use that
  script

- otherwise chose one of the following script (first hit wins = first file
  present in $SCRIPTS_DIR wins):
  - HOSTNAME.sh | HOSTNAME.master
  - GROUPNAME.sh | GROUPNAME.master
  - BASE_HOSTNAME.sh | BASE_HOSTNAME.master
  - IMAGENAME.sh | IMAGENAME.master

- pre/post install scripts are executed in the following order:
  - [0-9][0-9]all.*
  - [0-9][0-9]IMAGENAME.*
  - [0-9][0-9]BASE_HOSTNAME.*
  - [0-9][0-9]GROUPNAME.*
  - [0-9][0-9]HOSTNAME.*

This approach allows to define a hierarchical set of scripts that are executed
on the imaging client from the most global to the most local. In this way a more
local script (i.e. for a single host) can have override the effects of one or
more global scripts (i.e. group -> all and host -> group).

This approach preserves in part the backward compatibilty: everything defined in
prevsious versions is still executed in the same target of clients, but the
order of execution could change.

HINT: double check your scripts and take a look at the READMEs.