- fix: added UUID support for swap partitions
Unfortunately mkswap(8) doesn't allow to specify a UUID during the
initialization of the swap area: the UUID can be only generated randomly.

The problem is that the swap partitions will have a different UUID (respect to
the UUID specified in /etc/fstab cloned from the golden client) and it will be
not correctly mounted (swapon) at runtime in the imaged clients.

The workaround is to store the real device names (using the classical /dev
names) in /etc/fstab as comments and fix the UUID of the swap partitions using
the post-install script 10all.fix_swap_uuids.

Swap devices that don't use a UUID in /etc/fstab are considered as usual and
they will use the random UUID generated by mkswap.

Successfully tested with Ubuntu 7.04.