- allow to define hostname <-> imagename bindings via si_clusterconfig(8)
OK, if we can keep-in-sync files using the overrides and si_clusterconfig(8),
why to not use the same engine also to define hostname <-> imagename bindings?

The old approach via si_addclients is still available, and this is still the
most important. This means that if a matching symlink is found into the scripts
dir, this symlink will be used. Otherwise if there is no symlink for the
auto-installing client we'll use the info defined by si_clusterconfig(8).

The important thing is that it's fully backward compatible.