- fix: correctly handle double-click events in the Tk::HList object of   si_monitortk
Do not use the -command property of Tk::Hlist. Unfortunately the Tk::HList
object is totally buggy and it's not able to properly handle the double-click
events.

This means that under ceratin conditions the double-click handler will never be
called and the virtual console can be opened only using the <RETURN> key.

The workaround consists to manage the double-click event inside the
single-click handler, directly looking at the click-rate, evaluated by the
Time::HiRes features.