- allow to assign multiple overrides to a host group
Multiple overrides can be specified as a list of multiple XML tags:

<group>
   <name>GROUP_NAME</name>
   ...
   <override>OVERRIDE_NAME_1</override>
   <override>OVERRIDE_NAME_2</override>
   ...
   <override>OVERRIDE_NAME_N</override>
</group>

The OVERRIDE_NAME_1 ... OVERRIDE_NAME_N will be distributed preserving the same
order as they appear in the XML file. This means that in case of file overlaps
(more files in multiple overrides that should be distributed to the same target
filename) the first hit wins. In this case OVERRIDE_NAME_1 is the most
important and OVERRIDE_NAME_N is the least important.