- fix: force an initrd update at the end of the installation of SuSE images
This post-install script forces an initrd update in SuSE. This is needed because
the SuSE initrd includes only the block device modules detected in the golden
client.

If the imaged clients are equipped with a different hardware (typically a
different disk controller) they will not be able to come up after a reboot.

Bug reported by: keyesky <keyesky@gmail.com>