- added --imaging option to si_psh
This allows to concurrently execute commands on imaging clients that use the
installation parameter SSHD=y. With --imaging si_psh simply forks ssh with
the options "-o StrictHostKeyChecking=no -o UserKnownHostsFile=/dev/null" to
avoid the inclusion of the temporary ssh host key fingerprint in the
known_hosts of the image server (or the source host).