- added the "priority" attribute to host groups
If a node belongs to many host groups, in case of file overlaps, during the
distribution of the overrides, the group with the highest priority wins. A
small number means an high priority and vice versa.

If the priority is not defined or between groups that have the same priority
the importance is given by the alphabetical order of the group name.

In terms of DB language imaging that groups are sorted by "priority,name".