- fix: be sure to properly set the correct permissions bitmask in the   override directories.
According to MKDIR(2):

[...] the permissions of the created directory are (mode & ~umask & 0777).

If the umask used by root in the image server is not the commonly used 0022,
the override directories could be made with a permission bitmask different to
0755 that will be distributed in the root of the imaged clients.

A non-standard permission mask in the root can lead to serious problems, so
it's better to enforce the right bitmask directly using a chmod() just after
the mkdir().

The most reasonable solution here is to use the same permission mask of the
image, that is (by definition) the same mask used in the root of the golden
client.

Bug report and solution by: wtautz <wtautz@cs.uwaterloo.ca>