-fix: turn off the rsync quick checks when distributing overrides
Normally rsync will skip any files that are already the same size and have the
same modification timestamp. This can be a problem when multiple overrides are
distributed at the same time and there are more files with the same name, same
size and same timestamp (but different content) in the overrides and in the
image.

Under this rare (but potential) conditions a file is not correctly distributed
to the target clients. Typically this can happen when you have a script on the
image server that creates the same file name in multiple overrides and these
files are different, but have the same size (this can be quite common with
configuration files that must be distributed to different client sets).

The problem can be resolved adding the option "-I" to rsync only when
distributing overrides, in order to ignore the "name + size + timstamp" check
and enforce to always compare the content of the files. The only drawback of
this approach is a very small performance reduction when the overrides are
distributed to a running system (by si_pushoverrides or si_updateclient), but
it can resolve some troublesome issues.