- [PS3] fix: allow to build the BOEL kernel in Ubuntu distributions disabling   the stack protector features of gcc
The gcc 4.1 used in Ubuntu comes with the SSP (stack smashing protector)
build-in feature. Unfortunately the kernel can't be correctly compiled with
this option since SSP uses external symbols provided by libssp.so, that can't
be linked.

The workaround is to set the -fno-stack-protector option in all the CFLAGS of
the kernel Makefile.

See https://wiki.ubuntu.com/GccSsp for details.