- fix: updated rsync to 3.0.0pre4 into the BOEL initrd.img
This allows to resolve a rsync problem to set times on symlinks with kernels <
2.6.22. In this particular case rsync issues the system call lutimes() to
change the atime/mtime of a symlink, but the operating system returns an error
if it doesn't support it. The error is printed by rsync in the form:

rsync: failed to set times on "FILE": Function not implemented (38)

At the end of the imaging rsync exits with the error:

rsync error: some files could not be transferred (code 23) at main.c(980)

Since support for lutimes() was added to Linux in version 2.6.22, this could
happen if rsync is built on a system with kernel >= 2.6.22 and run it on a
system with kernel < 2.6.22.

See also:
- https://bugzilla.samba.org/show_bug.cgi?id=4977
- http://www.eu.kernel.org/pub/linux/kernel/v2.6/ChangeLog-2.6.22

Bug reported by: Rochus Schmid <rochus.schmid@rub.de>