- fix: netbootmond generates a wrong localboot file when deploying over a ssh tunnel
The IP of the imaged client is taken by netbootmond from the rsync log, but
when a SSH tunnel is used the remote peer address is always 127.0.0.1, so the
file 7F000001 in /tftpboot/pxelinux.cfg is created, instead of using the
correct IP address.

The solution is to simply append the IP address as a string to the "fake"
filename requested by the client to report the successful imaging completion.

If the appended IP address is found it will be used, otherwise we fallback to
the previous approach in order to preserve backward compatibility.

Bug reported by: Erich Focht <efocht@gmail.com>