- fix: correctly auto-detect a staging directory for the BitTorrent tarball   using the "df" implementation in BusyBox 1.8.1.
The df included in BusyBox 1.8.1 is not able to get the statfs() informations
of a mounted filesystem if a subdirectory is passed as argument, it requires 
the exact mountpoint instead. This fix extracts the mountpoints from the output
of df and use them to find a valid directory to stage the BitTorrent tarball.