- fix disk names in /boot/grub/device.map if it exists in the image
This is needed with System Configurator 2.2.12 (or greater). Starting from this
version System Configurator doesn't force the re-creation of device.map if it
already exists in the image, but this leads to failure if the imaged clients
have a different hardware respect to the golden client (i.e sda vs hda disks).

With System Configurator < 2.2.12 this is a no-op, because device.map is always
re-created by grub.