- fix: set correct disk names for systemconfigurator
The disk names in /etc/systemconfig/systemconfig.conf are changed in disk edits
procedure (in the same way as /etc/fstab, /boot/grub/device.map, etc.), but
this file is re-generated later, so the first change is lost.

To fix move the disk edit just before the systemconfigurator execution, after
the systemconfig.conf generation.