- improvements in si_lsimage
si_lsimage is now threaded, and it can retrieve details (-v) in parallel for
all the available images on the image server.

Moreover, now si_lsimage can accept one or more image names passed as argument
and print the details only for these selected images (or a warning for the
images that don't exist).