- fix: do not remove /dev/null on the build host
This workaround allows to save /dev/null when building SystemImager as root
(always discouraged!) on a host with old release of gcc.

Anyway, the problem is not a SystemImager issue, but a BusyBox issue, instead.
We hope Busybox will take care of this in the future.

See also: http://www.mail-archive.com/busybox@busybox.net/msg03258.html