- use hotplug to automatically load kernel modules in   autodetect_hardware_and_load_modules() with BOEL
With UYOK kernel modules are already loaded by hotplug in the early stage of
boot, since all the modules are included into the initrd.

With BOEL, the hotplug method is faster than pci-automod and it needs less
memory (this is very important for systems with a small amount of memory).

pci-automod is still needed by systemconfigurator and it'll be used at the
end of the imaging to create a valid /etc/systemconfig/hardware.lst before
running systemconfigurator. Also in terms of memory usage it's better to use
pci-automod at the end of imaging, in fact, at the very least, we can use the
swap partitions if it requires too much memory for small systems.