- fix: explicitly close rsync stdin when executing the rsync daemon by   si_prepareclient
When executed remotely via ssh rsync executed with --daemon tries to use stdin
as a socket to handle requests, instead of running in background and listening
for requests on the port given in the configuration file.

This fix explicitly closes the stdin of the spawned rsync daemon in
si_prepareclient, allowing to execute it remotely via ssh.

See also:
http://www.mail-archive.com/sisuite-users@lists.sourceforge.net/msg04519.html