added option --checksum (-c) to si_pcp and si_pushoverrides
This option allows to update only files with changed content (force checksum
usage for all the files on the sender). Receiver compares the checksums of
each file even if it exists and has the same size as the sender counterpart.

Mostly useful with -d (dry-run) to see which file will be altered by the
transfer.

Signed-off-by: Marco Innocenti <m.innocenti@cineca.it>